refactor(import): import/export maintainability + performance cleanups
Behavior-preserving cleanups to the Obsidian/Logseq import & export code,
addressing the deep-review findings in #1920 (maintainability) and #1921
(performance). No functional change except the deliberate i64->u64 count
type unification.

Maintainability (#1920):
- De-nest parse_frontmatter: lift BlockScalar to a module struct, replace the
  commit_block! macro with a fn, drop the unused chomp_strip field, and freeze
  the supported YAML subset (and rejected shapes) in a doc comment.
- Unify import counts on u64 (ImportResult + accumulators), removing the
  cast_unsigned juggling.
- Reference PAGE_LINK_RE/TAG_REF_RE via the canonical crate::cache path.
- Extract the shared YAML emit/parse helpers into commands::pages::markdown_yaml
  (the symmetric import/export quoting pair) with unit tests; lift FrontmatterRow.
- Document the cross-language [[wiki-link]] rule (Rust canonical) + add a
  Rust/TS regex parity test.
- Rename the broad import block-ref regex to PARENTHETICAL_REF_STRIP_RE.
- Batch the per-key frontmatter property_definitions lookup into one json_each
  query.

Performance (#1921):
- Pass the pre-fetched property declaration into set_property_in_tx (via a new
  with-declaration variant) to drop the double round-trip on the frontmatter path.
- strip_frontmatter returns Cow (zero-copy no-frontmatter case) and drops the
  per-line Vec allocation.
- Fast-path strip_block_refs and the inbound wiki-link rewrite to skip regex +
  allocation when the markers are absent.
- Document that IMPORT_CHUNK_BLOCKS bounds lock-hold time, not reproject cost.

Closes #1920
Closes #1921

+++ src-tauri/benches/import_bench.rs
@@ -140,7 +140,7 @@ fn bench_import_markdown_inner(c: &mut Criterion) {
                             .await
                             .unwrap();
 
-                            assert_eq!(result.blocks_created, n as i64);
+                            assert_eq!(result.blocks_created, n);
 
                             materializer.shutdown();
                             drop(dir);
